refact: #55 move src/apis -> src/utils/api
