+++ frontend/src/components/Header.test.jsx
@@ -10,7 +10,7 @@ import { mockStore } from '../mockStore';
 import rootReducer from '../modules';
 import 'jest-styled-components';
 import history from '../history';
-import axios from '../apis';
+import axios from '../utils/api';
 import Header from './Header';
 import NotificationDropdownList from './NotificationDropdownList';
 
@@ -20,7 +20,7 @@ jest.mock('../components/CustomQuestionModal', () => {
   });
 });
 
-jest.mock('../apis');
+jest.mock('../utils/api');
 
 jest.spyOn(axios, 'get').mockImplementation(() => {
   return new Promise((resolve) => {
