+++ frontend/src/modules/friend.js
@@ -1,4 +1,4 @@
-import axios from '../apis';
+import axios from '../utils/api';
 import { getNotifications } from './notification';
 
 const initialState = {
