+++ frontend/src/modules/like.js
@@ -1,4 +1,4 @@
-import axios from '../apis';
+import axios from '../utils/api';
 
 export const LIKE_POST_REQUEST = 'like/LIKE_POST_REQUEST';
 export const LIKE_POST_SUCCESS = 'like/LIKE_POST_SUCCESS';
