+++ frontend/src/modules/like.test.js
@@ -1,4 +1,4 @@
-import axios from '../apis';
+import axios from '../utils/api';
 import store from '../store';
 import { mockLike } from '../constants';
 import * as actionCreators from './like';
