+++ frontend/src/modules/post.test.js
@@ -1,4 +1,4 @@
-import axios from '../apis';
+import axios from '../utils/api';
 import {
   mockAnonymousFeed,
   mockArticle,
