+++ frontend/src/modules/question.test.js
@@ -1,5 +1,5 @@
 /* eslint-disable no-unused-vars */
-import axios from '../apis';
+import axios from '../utils/api';
 import store from '../store';
 import {
   mockQuestionFeed,
